+++ CHANGES.md
@@ -1,6 +1,9 @@
 Change Log
 ----------------
 - Version 3.0.5 (pending)
+    - Added right check to manual price override
+    - Added example test data to docs/testdata directory
+    - Added reprint button to Order print view
 
 - Version 3.0.4 (6/18/2018)
     - Show parental consent form checkbox in detail form